% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_dt.R
\name{as_pretty_dt}
\alias{as_pretty_dt}
\alias{cli_format.pretty_dt}
\title{Format a difftime}
\usage{
as_pretty_dt(x)

\method{cli_format}{pretty_dt}(x, style = NULL, ...)
}
\arguments{
\item{x}{A string representing a difftime.}

\item{style, ...}{Not used.}
}
\value{
An object of class \code{pretty_dt}.
}
\description{
Internal method used to print a \code{\link[=selenider_session]{selenider_session()}} object. Designed to
be used with \code{\link[prettyunits:pretty_dt]{prettyunits::pretty_dt()}}, \code{\link[prettyunits:pretty_ms]{prettyunits::pretty_ms()}} and
\code{\link[prettyunits:pretty_sec]{prettyunits::pretty_sec()}}.
}
\examples{
x <- as_pretty_dt(prettyunits::pretty_sec(10))

cli::cli_text("{.val x}")

}
\keyword{internal}
