% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session-options.R
\name{chromote_options}
\alias{chromote_options}
\alias{selenium_options}
\alias{selenium_server_options}
\alias{selenium_client_options}
\title{Driver options}
\usage{
chromote_options(
  headless = TRUE,
  user_agent = NULL,
  proxy_server = NULL,
  extra_args = NULL,
  parent = NULL,
  width = 992,
  height = 1323,
  targetId = NULL,
  wait_ = TRUE,
  auto_events = NULL
)

selenium_options(
  client_options = selenium_client_options(),
  server_options = selenium_server_options()
)

selenium_server_options(
  version = "latest",
  host = "localhost",
  port = 4444L,
  selenium_manager = TRUE,
  interactive = FALSE,
  stdout = NULL,
  stderr = NULL,
  verbose = FALSE,
  temp = TRUE,
  path = NULL,
  extra_args = c(),
  ...
)

selenium_client_options(
  port = 4444L,
  host = "localhost",
  verbose = FALSE,
  capabilities = NULL,
  request_body = NULL,
  timeout = 60
)
}
\arguments{
\item{headless}{Whether to run the browser in headless mode, meaning
that you won't actually be able to see the browser as you control it.
For debugging purposes and interactive use, it is often useful to set
this to \code{FALSE}.}

\item{user_agent}{A string containing a custom user agent to use for all
requests.}

\item{proxy_server}{A proxy server to use for all requests. Either a string
of the form \code{"HOST:PORT"}, or a list containing the \code{host} and \code{port}.
Optionally, \code{username} and \code{password} can be provided if the proxy server
requires authentication.}

\item{extra_args}{A character vector of arguments. In \code{chromote_options()},
these are passed to the Chrome process. In \code{selenium_client_options()},
these are passed to the Selenium server process.}

\item{parent}{The parent chromote session.}

\item{width, height, targetId, wait_, auto_events}{Passed into
\link[chromote:ChromoteSession]{chromote::ChromoteSession$new()}.}

\item{client_options}{A \code{\link[=selenium_client_options]{selenium_client_options()}} object.}

\item{server_options}{A \code{\link[=selenium_server_options]{selenium_server_options()}} object, or \code{NULL} if you
don't want one to be created.}

\item{version}{The version of Selenium server to use.}

\item{host}{The host to use.}

\item{port}{The port number to use.}

\item{selenium_manager, interactive, stdout, stderr, verbose, temp, path, ...}{Passed into \code{\link[selenium:selenium_server]{selenium::selenium_server()}}.}

\item{capabilities, request_body, timeout}{Passed into \link[selenium:SeleniumSession]{selenium::SeleniumSession$new()}.}
}
\description{
\code{chromote_options()} and \code{selenium_options()} return a list of options that
can be passed to the \code{options} argument of \code{selenider_session()}.

\code{chromote_options()} allows you to control the creation of a chromote driver
created using \link[chromote:ChromoteSession]{chromote::ChromoteSession$new()}.

\code{selenium_options()} allows you to control the creation of a selenium driver.

\code{selenium_server_options()} should be passed to the \code{server_options}
argument of \code{selenium_options()}, allowing you to control the creation of
the server using \code{\link[selenium:selenium_server]{selenium::selenium_server()}}.

\code{selenium_client_options()} should be passed to the \code{client_options} argument
of \code{selenium_options()}, allowing you to control the creation of a Selenium
client created using
\link[selenium:SeleniumSession]{selenium::SeleniumSession$new()}.
}
