% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{click}
\alias{click}
\title{Click an element}
\usage{
click(x, js = FALSE, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{js}{Whether to cick the element using JavaScript.}

\item{timeout}{How long to wait for the element to exist.}
}
\value{
\code{x}, invisibly
}
\description{
Clicks on an HTML element, either by simulating a mouse click or by triggering
the element's "click" event.
}
\examples{
session <- mock_selenider_session()

s(".class1") |>
  click()

}
\seealso{
Other actions: 
\code{\link{set_input}()}
}
\concept{actions}
