% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{close_session}
\alias{close_session}
\title{Close a session object}
\usage{
close_session(x = NULL)
}
\arguments{
\item{x}{A \code{selenider_session} object. If omitted, the global session object
will be closed.}
}
\value{
Nothing.
}
\description{
Shut down a session object, closing the browser and stopping the server.
}
\examples{
session <- mock_selenider_session()

close_session(session)

# Or:
close_session()

}
