% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{create_chromote_session}
\alias{create_chromote_session}
\alias{create_selenium_server}
\alias{create_selenium_client}
\alias{create_rselenium_client}
\title{Deprecated functions}
\usage{
create_chromote_session(parent = NULL, ...)

create_selenium_server(
  browser,
  version = "latest",
  driver_version = "latest",
  port = 4444L,
  quiet = TRUE,
  selenium_manager = TRUE,
  ...
)

create_selenium_client(browser, port = 4444L, host = "localhost", ...)

create_rselenium_client(browser, port = 4444L, ...)
}
\arguments{
\item{parent, ..., version, driver_version, port, quiet, host}{See the
documentation for \code{\link[=chromote_options]{chromote_options()}}, \code{\link[=selenium_options]{selenium_options()}},
\code{\link[=selenium_client_options]{selenium_client_options()}}, \code{\link[=wdman_server_options]{wdman_server_options()}},
\code{\link[=selenium_client_options]{selenium_client_options()}} and \code{\link[=rselenium_client_options]{rselenium_client_options()}} for details
about what these arguments mean.}

\item{browser}{The browser to use.}

\item{selenium_manager}{If this is \code{FALSE}, \code{\link[wdman:selenium]{wdman::selenium()}} will be used
instead of \code{\link[selenium:selenium_server]{selenium::selenium_server()}}. The equivalent of using
\code{\link[=wdman_server_options]{wdman_server_options()}} over \code{\link[=selenium_server_options]{selenium_server_options()}} in
\code{\link[=selenium_options]{selenium_options()}}.}
}
\value{
\code{create_chromote_session()} returns a \link[chromote:ChromoteSession]{chromote::ChromoteSession} object.

\code{create_selenium_server()} returns a \link[processx:process]{processx::process} or wdman
equivalent.

\code{create_selenium_client()} returns a \link[selenium:SeleniumSession]{selenium::SeleniumSession} object.

\code{create_rselenium_client()} returns an \link[RSelenium:remoteDriver-class]{RSelenium::remoteDriver} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions are deprecated and will be removed in a future release.
Use the \code{options} argument to \code{\link[=selenider_session]{selenider_session()}} instead. If you want
to manually create a chromote or selenium session, use
\link[chromote:ChromoteSession]{chromote::ChromoteSession}, \link[selenium:SeleniumSession]{selenium::SeleniumSession} and
\code{\link[selenium:selenium_server]{selenium::selenium_server()}} manually, since these functions
are only a thin wrapper around them.
}
