% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{create_rselenium_client}
\alias{create_rselenium_client}
\title{Start a Selenium session using RSelenium}
\usage{
create_rselenium_client(browser, port = 4444L, ...)
}
\arguments{
\item{browser}{The browser to use.}

\item{port}{The port to run RSelenium on.}

\item{...}{Other arguments to pass to RSelenium.}
}
\value{
An \link[RSelenium:remoteDriver-class]{RSelenium::remoteDriver} object. This can be passed into
\code{\link[=selenider_session]{selenider_session()}} in place of a \link[selenium:SeleniumSession]{selenium::SeleniumSession} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

We recommend using \code{\link[=create_selenium_client]{create_selenium_client()}} instead of this function,
as RSelenium is not compatible with newer versions of Selenium.
}
