% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_actions.R
\name{current_url}
\alias{current_url}
\title{Get the URL of the current page}
\usage{
current_url(session = NULL)
}
\arguments{
\item{session}{Optionally, a \code{selenider_session} object.}
}
\value{
A string: the current URL.
}
\description{
Get the full URL of the current page.
}
\examples{
\dontshow{if (selenider::selenider_available()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
session <- selenider_session()

open_url("https://r-project.org")

current_url()
\dontshow{\}) # examplesIf}
}
\seealso{
Other global actions: 
\code{\link{back}()},
\code{\link{execute_js_fn}()},
\code{\link{get_page_source}()},
\code{\link{open_url}()},
\code{\link{reload}()},
\code{\link{take_screenshot}()}
}
\concept{global actions}
