% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{elem_flatmap}
\alias{elem_flatmap}
\alias{element_list}
\title{Iterate over an element collection}
\usage{
elem_flatmap(x, .f, ...)

element_list(x, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_elements} object.}

\item{.f}{A function that takes a \code{selenider_element} and returns a
\code{selenider_element} or \code{selenider_elements} object.}

\item{...}{Passed into \code{.f}.}

\item{timeout}{How long to wait for \code{x} to exist while computing its length.}
}
\value{
\code{elem_flatmap()} returns a \code{selenider_elements} object.

\code{element_list()} returns a list of \code{selenider_element} objects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{elem_flatmap()} previously allowed you to apply a function to each
element in a collection in a lazy manner. This function is now deprecated,
as it did not work in all cases. Use \code{\link[=find_each_element]{find_each_element()}} and
\code{\link[=find_all_elements]{find_all_elements()}} instead for the simple case where you want to
select the children of a collection.

\code{element_list()} is a deprecated alias for \code{\link[=as.list.selenider_elements]{as.list.selenider_elements()}}.
}
