% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_actions.R
\name{get_page_source}
\alias{get_page_source}
\title{Read the HTML of a session}
\usage{
get_page_source(session = NULL, ...)
}
\arguments{
\item{session}{Optionally, a \code{selenider_session} object.}

\item{...}{Passed into \code{\link[=read_html.selenider_session]{read_html.selenider_session()}}}
}
\value{
An XML document, which can be operated on using \code{rvest}.
}
\description{
A shortcut for \code{xml2::read_html(get_session())}.
}
