% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_env.R
\name{get_session}
\alias{get_session}
\alias{set_session}
\title{Set or get the global selenider session}
\usage{
get_session()

set_session(session)
}
\arguments{
\item{session}{The session to set: a \code{\link[=selenider_session]{selenider_session()}} object.}
}
\value{
\code{get_session()} returns the current global \code{selenider_session} object, or
\code{NULL} if none have been created.
\code{set_session()} returns the previous global session.
}
\description{
\code{get_session()} gets the selenider session that is currently being used.
\code{set_session()} changes this session.
}
\examples{
session <- mock_session(set = FALSE)

get_session() # NULL

set_session(session)

get_session()

}
