% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{has_text}
\alias{has_text}
\alias{has_exact_text}
\title{Does an element contain a pattern?}
\usage{
has_text(x, text)

has_exact_text(x, text)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{text}{A string, used to test the element's inner text.}
}
\value{
A boolean value: TRUE or FALSE.
}
\description{
\code{has_text()} checks that an element's inner text contains a string, while
\code{has_exact_text()} checks that the inner text \emph{only} contains the string.
Both functions throw an error if the element does not exist in the DOM.
}
\details{
These functions do not implement a retry mechanism, and only test a condition
once. Use \code{\link[=html_expect]{html_expect()}} or \code{\link[=html_wait_until]{html_wait_until()}} to use these conditions in
tests.
}
\examples{
session <- mock_selenider_session()

has_text(s(".class1"), "Example")

}
\seealso{
Other conditions: 
\code{\link{is_enabled}()},
\code{\link{is_present}()},
\code{\link{is_visible}()}
}
\concept{conditions}
