% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{hover}
\alias{hover}
\title{Hover over an element}
\usage{
hover(x, js = FALSE, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{js}{Whether to hover the element using JavaScript.}

\item{timeout}{How long to wait for the element to exist.}
}
\value{
\code{x}, invisibly
}
\description{
Move the mouse over to an HTML element and hover over it, without actually
clicking or interacting with it.
}
\examples{
html <- "
<button onmouseover = settext()> </button>
<p class = 'text'></p>
"

js <- "
function settext() {
  document.getElementsByClassName('text').item(0).innerHTML = 'Button hovered!'
}
"

session <- minimal_selenider_session(html, js = js)

html_expect(s(".text"), has_exact_text(""))

s("button") |>
  hover()

html_expect(s(".text"), has_text("Button hovered!"))

}
\seealso{
Other actions: 
\code{\link{click}()},
\code{\link{scroll_to}()},
\code{\link{set_value}()},
\code{\link{submit}()}
}
\concept{actions}
