% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{html-actions}
\alias{html-actions}
\alias{click}
\alias{send_keys}
\alias{clear_input}
\alias{set_input}
\title{Perform actions on a selenider element.}
\usage{
click(x, js = FALSE, timeout = NULL)

send_keys(x, ..., timeout = NULL)

clear_input(x, timeout = NULL)

set_input(x, text, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{js}{Whether to cick the element using JavaScript.}

\item{timeout}{How long to wait for the element to exist.}

\item{...}{The keys to send to the element.}

\item{text}{The text to set the input to.}
}
\value{
\code{x}, invisibly.
}
\description{
These functions define various actions to be performed on a selenider
element.

\code{click()} clicks the element.

\code{send_keys()} sends a set of keys to an input element.

\code{clear_element()} clears the value of an input element.

\code{set_input()} sets the value of an input to a specified string.
}
\examples{
\dontshow{
# This allows `local_session()` to work when being sourced.
prev_options <- options(withr.hook_source = TRUE)
}
session <- mock_selenider_session()

s(".class1") |>
  click()
\dontshow{
options(prev_options)
}

}
