% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions_multiple.R
\name{html-conditions-multiple}
\alias{html-conditions-multiple}
\alias{has_length}
\alias{has_size}
\title{Conditions for selenider element collections}
\usage{
has_length(x, size)

has_size(x, size)
}
\arguments{
\item{x}{A \code{selenider_elements} object.}

\item{size}{The number of elements to expect.}
}
\value{
A boolean value: TRUE or FALSE.
}
\description{
These functions are predicates that test certain conditions on a
\code{selenider_elements} object, which are useful in conjunction with
\code{\link[=html_expect]{html_expect()}} and \code{\link[=html_wait_until]{html_wait_until()}}.

\code{has_length()}, or \code{has_size()} checks if an element collection contains a
certain number of elements.
}
\details{
These functions do not implement a retry mechanism, and only test a condition
once. Use \code{\link[=html_expect]{html_expect()}} or \code{\link[=html_wait_until]{html_wait_until()}} to use these conditions in
tests.
}
\examples{
\dontshow{
# This allows `local_session()` to work when being sourced.
prev_options <- options(withr.hook_source = TRUE)
}
session <- mock_selenider_session()

has_length(ss(".class1"), 2)
\dontshow{
options(prev_options)
}

}
