% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{html-properties}
\alias{html-properties}
\alias{html_name}
\alias{html_text}
\title{Get the properties of selenider elements}
\usage{
html_name(x, timeout = NULL)

html_text(x, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{timeout}{The time to wait for \code{x} to exist.}
}
\value{
A string.
}
\description{
Get various properties of an HTML element.

\code{html_name()} gets the tag name of an element.

\code{html_text()} gets the inner text of an element.
}
\examples{
session <- mock_selenider_session()

s(".class1") |>
  html_name()

}
