% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{html_ancestors}
\alias{html_ancestors}
\alias{html_parent}
\alias{html_siblings}
\alias{html_children}
\alias{html_descendants}
\title{Get the family elements of an element}
\usage{
html_ancestors(x)

html_parent(x)

html_siblings(x)

html_children(x)

html_descendants(x)
}
\arguments{
\item{x}{A \code{selenider_element} object.}
}
\value{
All functions return a \code{selenider_elements} object, except \code{html_parent()},
which returns a \code{selenider_element} object (since an element can only have one parent).
}
\description{
Find all elements with a certain relative position to an HTML element.

\code{html_ancestors()} selects every element which contains the current element
(children, grand-children, etc.).

\code{html_parent()} selects the element that contains the current element.

\code{html_siblings()} selects every element which has the same parent as the current
element.

\code{html_children()} selects every element which is connected to and directly below
the current element.

\code{html_descendants()} selects every element that is contained by the current element.
The current element does not have to be a direct parent, but must be some type of
ancestor.
}
\seealso{
\itemize{
\item \url{http://web.simmons.edu/~grovesd/comm244/notes/week4/document-tree} for a simple
and visual explanation of the document tree.
}
}
