% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{html_attr}
\alias{html_attr}
\title{Get element attribute}
\usage{
html_attr(x, name, default = NA_character_, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{name}{The name of the attribute to get; a string.}

\item{default}{The default value to use if the attribute does not exist in
the element.}

\item{timeout}{The time to wait for \code{x} to exist.}
}
\value{
A character vector.
}
\description{
Get an attribute of a \code{selenider_element} object.
}
\examples{
session <- mock_selenider_session()

s(".class1") |>
  html_attr("href")

}
\seealso{
Other properties: 
\code{\link{html_name}()},
\code{\link{html_text}()}
}
\concept{properties}
