% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{html_css_property}
\alias{html_css_property}
\title{Get a CSS property of an element}
\usage{
html_css_property(x, name, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{name}{The name of the CSS property to get.}

\item{timeout}{The time to wait for \code{x} to exist.}
}
\value{
A string, or \code{NA} if the property does not exist.
}
\description{
Get a CSS property of an element (e.g. \code{"background-color"}).
}
\examples{
session <- mock_selenider_session()

s(".class1") |>
  html_css_property("background-color")

}
\seealso{
Other properties: 
\code{\link{html_attr}()},
\code{\link{html_name}()},
\code{\link{html_text}()}
}
\concept{properties}
