% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{html_text}
\alias{html_text}
\alias{html_text.selenider_element}
\title{Get element text}
\usage{
html_text(x, ...)

\method{html_text}{selenider_element}(x, timeout = NULL, ...)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{...}{Not used.}

\item{timeout}{The time to wait for \code{x} to exist.}
}
\value{
A string
}
\description{
Get the inner text of a \code{selenider_element} object.
}
\examples{
\dontshow{if (selenider_available(online = FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
html <- "
<p>Example text</p>
"

session <- minimal_selenider_session(html)

s("p") |>
  html_text()

\dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other properties: 
\code{\link{html_attr}()},
\code{\link{html_css_property}()},
\code{\link{html_name}()}
}
\concept{properties}
