% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{is_present}
\alias{is_present}
\alias{is_in_dom}
\alias{is_absent}
\title{Does an element exist?}
\usage{
is_present(x)

is_in_dom(x)

is_absent(x)
}
\arguments{
\item{x}{A \code{selenider_element} object.}
}
\value{
A boolean value: TRUE or FALSE.
}
\description{
\code{is_present()} and \code{is_in_dom()} checks if an element is present on the page,
while \code{is_missing()} and \code{is_absent()} checks the opposite.
}
\details{
These functions do not implement a retry mechanism, and only test a condition
once. Use \code{\link[=html_expect]{html_expect()}} or \code{\link[=html_wait_until]{html_wait_until()}} to use these conditions in
tests.
}
\examples{
session <- mock_selenider_session()

is_present(s(".class1"))

}
\seealso{
Other conditions: 
\code{\link{has_attr}()},
\code{\link{has_css_property}()},
\code{\link{has_name}()},
\code{\link{has_text}()},
\code{\link{is_enabled}()},
\code{\link{is_visible}()}
}
\concept{conditions}
