% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{is_visible}
\alias{is_visible}
\alias{is_displayed}
\alias{is_hidden}
\alias{is_invisible}
\title{Is an element visible?}
\usage{
is_visible(x)

is_displayed(x)

is_hidden(x)

is_invisible(x)
}
\arguments{
\item{x}{A \code{selenider_element} object.}
}
\value{
A boolean value: TRUE or FALSE.
}
\description{
\code{is_visible()} and \code{is_displayed()} checks that an element can be seen on the
page, while \code{is_invisible()} and \code{is_hidden()} checks the opposite. All
functions throw an error if the element is not in the DOM.
}
\details{
These functions do not implement a retry mechanism, and only test a condition
once. Use \code{\link[=html_expect]{html_expect()}} or \code{\link[=html_wait_until]{html_wait_until()}} to use these conditions in
tests.
}
\examples{
session <- mock_selenider_session()

is_visible(s(".class1"))

}
\seealso{
Other conditions: 
\code{\link{has_text}()},
\code{\link{is_enabled}()},
\code{\link{is_present}()}
}
\concept{conditions}
