% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\docType{data}
\name{keys}
\alias{keys}
\title{Special keys}
\format{
An object of class \code{list} of length 28.
}
\usage{
keys
}
\description{
List of special keys, for use with \code{\link[=send_keys]{send_keys()}}.
}
\examples{
keys$backspace

}
\keyword{datasets}
