% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_selenider_session.R
\name{minimal_selenider_session}
\alias{minimal_selenider_session}
\title{Create a session with custom HTML}
\usage{
minimal_selenider_session(html, js = NULL, ..., .env = rlang::caller_env())
}
\arguments{
\item{html}{A string to use as HTML.}

\item{js}{A string (or \code{NULL}) to use as JavaScript.}

\item{...}{Passed into \code{\link[=selenider_session]{selenider_session()}}.}

\item{.env}{The environment in which the session will be used.}
}
\value{
A \code{selenider_session} object.
}
\description{
Create a \code{selenider_session} using custom HTML/JavaScript. Similar to
\code{\link[rvest:minimal_html]{rvest::minimal_html()}}.
}
