% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_actions.R
\name{open_url}
\alias{open_url}
\title{Open a URL}
\usage{
open_url(url, session = NULL)
}
\arguments{
\item{url}{The URL to navigate to: a string.}

\item{session}{A `selenider_session object. If not specified, the global
session object (the result of \code{\link[=get_session]{get_session()}}) is used.}
}
\value{
The session object, invisibly.
}
\description{
Navigate the browser to specified URL, waiting until the page is considered
open before finishing.
}
\examples{
session <- mock_selenider_session()

open_url("https://www.google.com")

# Or:
open_url(session = session, "https://www.google.com")

}
\seealso{
Other global actions: 
\code{\link{back}()},
\code{\link{reload}()},
\code{\link{take_screenshot}()}
}
\concept{global actions}
