% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_lazy.R
\name{print_lazy}
\alias{print_lazy}
\alias{print_lazy.selenider_element}
\alias{print_lazy.selenider_elements}
\title{Print an element without fetching it}
\usage{
print_lazy(x, ...)

\method{print_lazy}{selenider_element}(x, ...)

\method{print_lazy}{selenider_elements}(x, ...)
}
\arguments{
\item{x}{A \code{selenider_element} or \code{selenider_elements} object.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Display a summary of the steps needed to reach an element. This function
is deprecated, as it is not useful for most users.
}
