% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_actions.R
\name{reload}
\alias{reload}
\alias{refresh}
\title{Reload the current page}
\usage{
reload(session = NULL)

refresh(session = NULL)
}
\arguments{
\item{session}{A `selenider_session object. If not specified, the global
session object (the result of \code{\link[=get_session]{get_session()}}) is used.}
}
\value{
The session object, invisibly.
}
\description{
\code{reload()} and \code{refresh()} both reload the current page.
}
\examples{
session <- mock_selenider_session()

open_url("https://www.google.com")

reload()

}
\seealso{
Other global actions: 
\code{\link{back}()},
\code{\link{open_url}()},
\code{\link{take_screenshot}()}
}
\concept{global actions}
