% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{scroll_to}
\alias{scroll_to}
\title{Scroll to an element}
\usage{
scroll_to(x, js = FALSE, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{js}{Whether to scroll to the element using JavaScript.}

\item{timeout}{How long to wait for the element to exist.}
}
\value{
\code{x}, invisibly
}
\description{
Scrolls to an HTML element.
}
\examples{
html <- "
<div style = 'height:100\%; min-height:100vh'></div>
<button onclick='checkScrolled()'></button>
<p>Scroll down to find me!</p>
"

js <- "
function checkScrolled() {
  let element = document.getElementsByTagName('p').item(0);
  let rect = element.getBoundingClientRect();
  // If paragraph is in view
  if (rect.bottom <= (window.innerHeight || document.documentElement.clientHeight)) {
    element.innerText = 'You found me!';
  }
}
"

session <- minimal_selenider_session(html, js = js)

s("p") |>
  scroll_to()

s("button") |>
  click()

html_expect(s("p"), has_text("You found me!"))

}
\seealso{
Other actions: 
\code{\link{click}()},
\code{\link{hover}()},
\code{\link{set_value}()},
\code{\link{submit}()}
}
\concept{actions}
