% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_actions.R
\name{open_url}
\alias{open_url}
\alias{back}
\alias{forward}
\alias{reload}
\alias{refresh}
\alias{take_screenshot}
\title{Actions on the current webpage}
\usage{
open_url(url, session = NULL)

back(session = NULL)

forward(session = NULL)

reload(session = NULL)

refresh(session = NULL)

take_screenshot(session, file = NULL)
}
\arguments{
\item{url}{The URL to navigate to.}

\item{session}{A `selenider_session object. If not specified, the global
session object (the result of \code{\link[=get_session]{get_session()}}) is used.}

\item{file}{The file path to save the screenshot to.}
}
\value{
The session object, invisibly.
}
\description{
These functions perform certain actions that are specific to the current
session, rather than a specific element. Each can be used with or without
explicitly specifying the session object.

\code{open_url()} navigates the webpage to a specified URL, waiting until the
page is considered open.

\code{back()} navigates to the previously opened URL, or the previously opened
page in your browsing history.

\code{forward()} reverses the action of \code{back()}, going to the next page in your
browsing history.

\code{reload()}, or \code{refresh()}, reloads the current page.

\code{take_screenshot()} takes a screenshot of the page currently being viewed,
saving it to a file.
}
\details{
Arguments to all functions are processed in a way such that enable the
session argument to be omitted.
}
\examples{
session <- mock_selenider_session()

open_url(session, "https://www.google.com")

# Or:
open_url("https://www.google.com")

}
