% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{selenider-config}
\alias{selenider-config}
\title{Selenider options}
\description{
\code{selenider} has a few options, allowing you to specify the session and
browser to use without having to tell \code{\link[=selenider_session]{selenider_session()}} this information
every time.
\itemize{
\item \code{selenider.session} - The package to use as a backend: either "chromote",
"selenium" or "rselenium".
\item \code{selenider.browser} - The name of the browser to run the session in; one
of "chrome", "firefox", "edge", "safari", or another valid browser name.
}
}
