% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenider-package.R
\docType{package}
\name{selenider-package}
\alias{selenider}
\alias{selenider-package}
\title{selenider: Concise, lazy and asynchronous wrapper for 'RSelenium'}
\description{
A wrapper for 'RSelenium' that is much easier to use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ashbythorpe/selenider}
  \item \url{https://ashbythorpe.github.io/selenider/}
  \item Report bugs at \url{https://github.com/ashbythorpe/selenider/issues}
}

}
\author{
\strong{Maintainer}: Ashby Thorpe \email{ashbythorpe@gmail.com} (\href{https://orcid.org/0000-0003-3106-099X}{ORCID})

}
\keyword{internal}
