% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selenider_cleanup}
\alias{selenider_cleanup}
\title{Cleanup after an example}
\usage{
selenider_cleanup(close_session = TRUE, env = rlang::caller_env())
}
\arguments{
\item{close_session}{Whether there is still a local session that needs to be
closed (via \code{\link[withr:defer]{withr::deferred_run()}}). This is usually the case.}

\item{env}{The environment to get the local session.}
}
\description{
Clean up after a selenider example, making sure all environment variables are
reset and no connections are left open.
}
\keyword{internal}
