% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvest.R
\name{selenider_prefer}
\alias{selenider_prefer}
\alias{read_html.selenider_session}
\alias{read_html.selenider_element}
\alias{html_table.selenider_session}
\alias{html_table.selenider_element}
\title{Compatibility with rvest}
\usage{
selenider_prefer(quiet = TRUE)

read_html.selenider_session(
  x,
  encoding = "",
  ...,
  options = c("RECOVER", "NOERROR", "NOBLANKS")
)

read_html.selenider_element(
  x,
  encoding = "",
  timeout = NULL,
  outer = TRUE,
  ...,
  options = c("RECOVER", "NOERROR", "NOBLANKS")
)

html_table.selenider_session(
  x,
  header = NA,
  trim = TRUE,
  fill = deprecated(),
  dec = ".",
  na.strings = "NA",
  convert = TRUE
)

html_table.selenider_element(
  x,
  header = NA,
  trim = TRUE,
  fill = deprecated(),
  dec = ".",
  na.strings = "NA",
  convert = TRUE
)
}
\arguments{
\item{quiet}{If \code{TRUE}, output from \code{\link[conflicted:conflict_prefer]{conflicted::conflict_prefer()}} will
be suppressed.}

\item{x}{A \code{selenider_session}/\code{selenider_element} object.}

\item{encoding, ..., options}{Passed into \code{\link[xml2:read_xml]{xml2::read_html()}}.}

\item{timeout}{How long to wait for \code{x} to exist in the DOM before throwing
an error.}

\item{outer}{Whether to read the inner (all children of the current element) or
outer (including the element itself) HTML of \code{x}.}

\item{header, trim, fill, dec, na.strings, convert}{Passed into \code{\link[rvest:html_table]{rvest::html_table()}}.}
}
\value{
\code{read_html()} returns an XML document.
\code{html_table()} returns a tibble or list of tibbles (when applied to a page).
}
\description{
These functions ensure compatibility with \code{rvest} and \code{xml2}.

\code{selenider_prefer()} ensures that all functions common to \code{selenider} and
\code{rvest} (e.g. \code{html_attr()}) work on both selenider elements and rvest node
objects. \code{conflicted} is used to resolve function conflicts.

selenider allows \code{\link[xml2:read_xml]{xml2::read_html()}} and \code{\link[rvest:html_table]{rvest::html_table()}} to work
on \code{selenider_element} and \code{selenider_session} objects.
}
\examples{
\dontshow{if (selenider_available(online = FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(rvest)
selenider_prefer(quiet = FALSE)

html <- "
<div>
<p>Example text</p>
</div>

<table>
  <tr>
    <th>Column 1</th>
    <th>Column 2</th>
  </tr>
  <tr>
    <td>10</td>
    <td>Data</td>
  </tr>
</table>
"

session <- minimal_selenider_session(html)

read_html(session)
read_html(s("div"))

html_table(s("table"))

\dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
}
\dontshow{\}) # examplesIf}
}
