% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{set_input}
\alias{set_input}
\alias{send_keys}
\alias{clear_input}
\title{Set the value of an input}
\usage{
set_input(x, text, timeout = NULL)

send_keys(x, ..., timeout = NULL)

clear_input(x, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{text}{A string to set the input}

\item{timeout}{How long to wait for the element to exist.}

\item{...}{A set of inputs to send to \code{x}.}
}
\value{
\code{x}, invisibly
}
\description{
\code{set_input()} sets the value of an HTML input element to a string.
}
\details{
\code{send_keys()} sends a set of inputs to an element.

\code{clear_input()} sets the value of an HTML element to \code{""}, removing any
existing content.
}
\examples{
session <- mock_selenider_session()

s(".class1") |>
  set_input("my text")

}
\seealso{
Other actions: 
\code{\link{click}()}
}
\concept{actions}
