% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{set_value}
\alias{set_value}
\alias{send_keys}
\alias{clear_value}
\title{Set the value of an input}
\usage{
set_value(x, text, timeout = NULL)

send_keys(x, ..., modifiers = NULL, timeout = NULL)

clear_value(x, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{text}{A string to set the input}

\item{timeout}{How long to wait for the element to exist.}

\item{...}{A set of inputs to send to \code{x}.}

\item{modifiers}{A character vector; one or more of "shift", "ctrl", "alt", and "command"/meta".}
}
\value{
\code{x}, invisibly
}
\description{
\code{set_value()} sets the value of an HTML input element to a string.
}
\details{
\code{send_keys()} sends a set of inputs to an element.

\code{clear_value()} sets the value of an HTML element to \code{""}, removing any
existing content.
}
\examples{
session <- mock_selenider_session()

s(".class1") |>
  set_value("my text")

}
\seealso{
Other actions: 
\code{\link{click}()},
\code{\link{hover}()},
\code{\link{scroll_to}()},
\code{\link{submit}()}
}
\concept{actions}
