% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{submit}
\alias{submit}
\title{Submit an element}
\usage{
submit(x, js = FALSE, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{js}{Whether to submit the form using JavaScript.}

\item{timeout}{How long to wait for the element to exist.}
}
\value{
\code{x}, invisibly
}
\description{
If an element is an ancestor of a form, submits the form.
Works by walking up the DOM, checking each ancestor element until
the element is a \if{html}{\out{<form>}} element, which it then submits. If such
an element does not exist, an error is thrown.
}
\examples{
session <- mock_selenider_session()

s(".class1") |>
  submit()

}
\seealso{
Other actions: 
\code{\link{click}()},
\code{\link{hover}()},
\code{\link{scroll_to}()},
\code{\link{set_value}()}
}
\concept{actions}
