% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_actions.R
\name{take_screenshot}
\alias{take_screenshot}
\title{Take a screenshot of the current page}
\usage{
take_screenshot(file = NULL, session = NULL)
}
\arguments{
\item{file}{The file path to save the screenshot to.}

\item{session}{A `selenider_session object. If not specified, the global
session object (the result of \code{\link[=get_session]{get_session()}}) is used.}
}
\value{
The session object, invisibly.
}
\description{
Take a screenshot of the current session state, saving this image to a file.
}
\examples{
session <- mock_selenider_session()

open_url("https://www.google.com")

file_path <- tempfile(fileext = "jpeg")

take_screenshot(file_path)

}
\seealso{
Other global actions: 
\code{\link{back}()},
\code{\link{open_url}()},
\code{\link{reload}()}
}
\concept{global actions}
