% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{click}
\alias{click}
\alias{double_click}
\alias{right_click}
\title{Click an element}
\usage{
click(x, js = FALSE, timeout = NULL)

double_click(x, js = FALSE, timeout = NULL)

right_click(x, js = FALSE, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{js}{Whether to click the element using JavaScript.}

\item{timeout}{How long to wait for the element to exist.}
}
\value{
\code{x}, invisibly
}
\description{
Clicks on an HTML element, either by simulating a mouse click or by triggering
the element's "click" event.

\code{click()} left clicks on the element, \code{double_click()} left clicks on the
element two times in a short period of time, while \code{right_click()} right
clicks on an element, opening its context menu.
}
\examples{
html <- "
<button onclick = hidetext() oncontextmenu = showtext()></button>
<p id = 'texttohide'>Hello!</p>
"

js <- "
function hidetext() {
  document.getElementById('texttohide').style.display = 'none'
}

function showtext() {
  document.getElementById('texttohide').style.display = 'block'
}
"

session <- minimal_selenider_session(html, js = js)

html_expect(s("p"), is_visible)

s("button") |>
  click()

html_expect(s("p"), is_invisible)

s("button") |>
  right_click()

html_expect(s("p"), is_visible)

}
\seealso{
Other actions: 
\code{\link{hover}()},
\code{\link{scroll_to}()},
\code{\link{set_value}()},
\code{\link{submit}()}
}
\concept{actions}
