% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{close_session}
\alias{close_session}
\title{Close a session object}
\usage{
close_session(x = NULL)
}
\arguments{
\item{x}{A \code{selenider_session} object. If omitted, the global session object
will be closed.}
}
\value{
Nothing.
}
\description{
Shut down a session object, closing the browser and stopping the server.
}
\examples{
session <- mock_selenider_session(local = FALSE)

close_session(session)

# Reopen the session, this time letting it be set locally.
session <- mock_selenider_session()

# We don't have to specify the session if it is set locally.
close_session()

# Since we already closed `session`, we don't need the deferred events to
# run
withr::deferred_clear()

}
