% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_actual_element.R
\name{get_actual_webelement}
\alias{get_actual_webelement}
\alias{get_actual_webelements}
\title{Get the webElement associated with a selenider element}
\usage{
get_actual_webelement(x, timeout = NULL)

get_actual_webelements(x, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} or \code{selenider_elements} object, produced by
\code{\link[=html_element]{html_element()}} / \code{\link[=html_elements]{html_elements()}}}

\item{timeout}{The timeout to use while asserting that the item exists. If
NULL, the timeout of the \code{selenider_element} will be used.}
}
\value{
A \link[RSelenium:webElement-class]{RSelenium::webElement} object, or a list of such objects.
}
\description{
Turn a lazy selenium element or element collection into a
\link[RSelenium:webElement-class]{RSelenium::webElement}. Use this to perform certain actions on the element
that are not implemented in selenider (e.g. getElementLocation())

\code{get_actual_webelement()} turns a \code{selenider_element} object into a
\link[RSelenium:webElement-class]{RSelenium::webElement} object. The function will wait for the object to
exist in the DOM.

\code{get_actual_webelements()} turns a \code{selenider_elements} object into a list
of \link[RSelenium:webElement-class]{RSelenium::webElement} objects.
}
\examples{
\dontshow{if (selenider_available(online = FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
html <- "
<div>
<p>Text</p>
<p>More text</p>
</div>
"

session <- minimal_selenider_session(html)

elem <- s("div") |>
  get_actual_webelement()

# The webDriver/ChromoteSession can be accessed using session$driver
driver <- session$driver

if (inherits(driver, "ChromoteSession")) {
  driver$DOM$getBoxModel(backendNodeId = elem)
} else {
  elem$getElementLocation()
}

elems <- ss("p") |>
  get_actual_webelements()

if (inherits(driver, "ChromoteSession")) {
  driver$DOM$describeNode(backendNodeId = elems[[1]])
} else {
  elems[[1]]$describeElement()
}

\dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
}
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \code{\link[=s]{s()}}, \code{\link[=ss]{ss()}}, \code{\link[=html_element]{html_element()}} and \code{\link[=html_elements]{html_elements()}} to select selenider
elements
\item The documentation for \code{\link[RSelenium:webElement-class]{RSelenium::webElement()}} to see the things you can
do with them.
}
}
