% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{exists}
\alias{exists}
\alias{is_present}
\alias{is_in_dom}
\alias{is_missing}
\alias{is_absent}
\alias{is_visible}
\alias{is_displayed}
\alias{is_hidden}
\alias{is_invisible}
\alias{is_enabled}
\alias{is_disabled}
\title{Conditions for selenider elements}
\usage{
exists(x)

is_present(x)

is_in_dom(x)

is_missing(x)

is_absent(x)

is_visible(x)

is_displayed(x)

is_hidden(x)

is_invisible(x)

is_enabled(x)

is_disabled(x)
}
\arguments{
\item{x}{A \code{selenider_element} object.}
}
\value{
A boolean value: TRUE or FALSE.
}
\description{
These functions are predicates that test certain conditions on a
\code{selenider_element} object, which are useful in conjunction with
\code{\link[=html_expect]{html_expect()}} and \code{\link[=html_wait_until]{html_wait_until()}}.

\code{exists()}, \code{is_present()} and \code{is_in_dom()} checks if an element is present
on the page, while \code{is_missing()} and \code{is_absent()} checks the opposite.

\code{is_visible()} and \code{is_displayed()} checks that an element can be seen on the
page, while \code{is_invisible()} and \code{is_hidden()} checks the opposite.

\code{is_disabled()} checks that an element has the \code{disabled} attribute set to
\code{TRUE}, while \code{is_enabled()} checks that it does not.
}
\details{
These functions do not implement a retry mechanism, and only test a condition
once. Use \code{\link[=html_expect]{html_expect()}} or \code{\link[=html_wait_until]{html_wait_until()}} to use these conditions in
tests.
}
\examples{
session <- mock_selenider_session()

exists(s(".class1"))

}
