% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children.R
\name{html_ancestors}
\alias{html_ancestors}
\alias{html_parent}
\alias{html_siblings}
\alias{html_children}
\alias{html_descendants}
\title{Get the DOM family of an element}
\usage{
html_ancestors(x)

html_parent(x)

html_siblings(x)

html_children(x)

html_descendants(x)
}
\arguments{
\item{x}{A \code{selenider_element} object.}
}
\value{
All functions return a \code{selenider_elements} object, except \code{html_parent()},
which returns a \code{selenider_element} object (since an element can only have one parent).
}
\description{
Find all elements with a certain relative position to an HTML element.

\code{html_ancestors()} selects every element which contains the current element
(children, grand-children, etc.).

\code{html_parent()} selects the element that contains the current element.

\code{html_siblings()} selects every element which has the same parent as the current
element.

\code{html_children()} selects every element which is connected to and directly below
the current element.

\code{html_descendants()} selects every element that is contained by the current element.
The current element does not have to be a direct parent, but must be some type of
ancestor.
}
\details{
All functions except \code{html_children()} and \code{html_descendants()} use XPath selectors,
so may be slow, especially when using \code{chromote} as a backend.
}
\examples{
\dontshow{if (selenider_available(online = FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
html <- "
<html>
<body>
  <div>
    <div id='current'>
      <p></p>
      <div>
        <p></p>
        <br>
      </div>
    </div>
    <div></div>
    <p></p>
  </div>
</body>
</html>
"

session <- minimal_selenider_session(html)

current <- s("#current")

# Get all the names of an element collection
html_names <- function(x) {
  x |>
    as.list() |>
    vapply(html_name, FUN.VALUE = character(1))
}

current |>
  html_ancestors() |>
  html_expect(has_length(3)) |>
  html_names() # html, div, body

current |>
  html_parent() |>
  html_name() # div

current |>
  html_siblings() |>
  html_expect(has_length(2)) |>
  html_names() # div, p

current |>
  html_children() |>
  html_expect(has_length(2)) |>
  html_names() # p, div

current |>
  html_descendants() |>
  html_expect(has_length(4)) |>
  html_names() # p, div, p, br

\dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
}
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \url{http://web.simmons.edu/~grovesd/comm244/notes/week4/document-tree} for a simple
and visual explanation of the document tree.
\item \code{\link[=html_element]{html_element()}} and \code{\link[=html_elements]{html_elements()}} for other ways of selecting elements. These
functions allow you to select ancestors using one or more conditions (e.g. CSS
selectors).
\item \code{\link[=html_filter]{html_filter()}} and \code{\link[=html_find]{html_find()}} for filtering element collections.
}
}
