% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_elements.R
\name{html_elements}
\alias{html_elements}
\alias{html_elements.selenider_session}
\alias{html_elements.selenider_element}
\title{Get multiple HTML elements}
\usage{
html_elements(x, ...)

\method{html_elements}{selenider_session}(
  x,
  css = NULL,
  xpath = NULL,
  id = NULL,
  class_name = NULL,
  name = NULL,
  link_text = NULL,
  ...
)

\method{html_elements}{selenider_element}(
  x,
  css = NULL,
  xpath = NULL,
  id = NULL,
  class_name = NULL,
  name = NULL,
  link_text = NULL,
  ...
)
}
\arguments{
\item{x}{A selenider session or element.}

\item{...}{Arguments passed to methods.}

\item{css}{A css selector.}

\item{xpath}{An XPath.}

\item{id}{The id of the element you want to select.}

\item{class_name}{The class name of the element you want to select.}

\item{name}{The name attribute of the element you want to select.}

\item{link_text}{The link text of the link element that you would like to
select.}
}
\value{
A \code{selenider_elements} object.
}
\description{
Find every available HTML element using a CSS selector, an XPath, or a
variety of other methods.
}
\details{
If more than one method is used to select an element (e.g. \code{css} and
\code{xpath}), the first element which satisfies every condition will be found.
}
\examples{
\dontshow{if (selenider_available(online = FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
html <- "
<div id='outer-div'>
  <div>
    <p>Text 1</p>
    <p>Text 2</p>
    <p>Text 3</p>
  </div>
</div>

<div></div>
"

session <- minimal_selenider_session(html)

session |>
  html_elements("div")

# Or:
ss("div")

session |>
  html_element("#outer-div") |>
  html_elements("p")
  
# The above can be shortened to:
s("#outer-div") |>
  html_elements("p")

\dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
}
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \code{\link[=ss]{ss()}} to quickly select multiple elements without specifying the session.
\item \code{\link[=html_element]{html_element()}} to select multiple elements.
\item \code{\link[=selenider_session]{selenider_session()}} to begin a session.
\item \code{\link[=html_children]{html_children()}} and family to select elements using their relative position
in the DOM.
\item \code{\link[=html_filter]{html_filter()}} and \code{\link[=html_find]{html_find()}} for filtering element collections.
}
}
