% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal.R
\name{html_equal}
\alias{html_equal}
\alias{==.selenider_element}
\title{Are two elements equivalent?}
\usage{
html_equal(x, y, timeout = NULL)

\method{==}{selenider_element}(e1, e2)
}
\arguments{
\item{x, y, e1, e2}{\code{selenider_element} objects to compare.}

\item{timeout}{How long to wait for the elements to be present.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Checks if two \code{selenider_element} objects point to the
same element on the page. \code{html_equal()} is equivalent to
using \code{==}, but allows you to specify a timeout value if
needed.
}
