% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_flatten.R
\name{html_flatten}
\alias{html_flatten}
\alias{c.selenider_element}
\alias{c.selenider_elements}
\title{Combine multiple HTML elements}
\usage{
html_flatten(...)

\method{c}{selenider_element}(...)

\method{c}{selenider_elements}(...)
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> \code{selenider_element} or
\code{selenider_elements} objects to be combined, or lists of such objects.}
}
\value{
A \code{selenider_elements} object.
}
\description{
Combine a set of \code{selenider_element}/\code{selenider_elements} objects
into a single \code{selenider_elements} object, allowing you to
perform actions on them at once. \code{c()} and \code{html_flatten()} do the same
thing, but \code{html_flatten()} works on lists of \code{selenider_element}/\code{selenider_elements}
objects.
}
\examples{
session <- mock_selenider_session()

element <- s(".class1")

collection <- ss(".class2")

html_flatten(element, collection)

c(collection, element)

html_flatten(list(element, collection))

}
