% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties-multiple.R
\name{html_size}
\alias{html_size}
\alias{length.html_elements}
\title{Get the number of elements in a collection}
\usage{
html_size(x, timeout = NULL)

\method{length}{html_elements}(x)
}
\arguments{
\item{x}{A \code{selenider_elements} object.}

\item{timeout}{}
}
\value{
An integer representing the number of elements in the collection.
}
\description{
Get the number of elements in a HTML element collection, waiting for the
parent elements (if any) to exist before returning a value.

\code{length()} and \code{html_size()} can be used interchangeably, the only
difference being that \code{html_size()} allows you to specify a timeout.
}
\examples{
session <- mock_selenider_session()

ss(".class1") |>
  length()

}
\concept{collection properties}
