% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{html_text}
\alias{html_text}
\title{Get element text}
\usage{
html_text(x, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{timeout}{The time to wait for \code{x} to exist.}
}
\value{
A string
}
\description{
Get the inner text of a \code{selenider_element} object.
}
\examples{
session <- mock_selenider_session()

s(".class1") |>
  html_text()

}
\seealso{
Other properties: 
\code{\link{html_attr}()},
\code{\link{html_name}()}
}
\concept{properties}
