% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\docType{data}
\name{keys}
\alias{keys}
\title{Special keys}
\format{
A list containing \code{selenider_key} objects.
}
\usage{
keys
}
\description{
List of special keys, for use with \code{\link[=elem_send_keys]{elem_send_keys()}}.
}
\examples{
keys$backspace
}
\keyword{datasets}
