% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{scroll_to}
\alias{scroll_to}
\title{Scroll to an element}
\usage{
scroll_to(x, js = FALSE, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object.}

\item{js}{Whether to scroll to the element using JavaScript.}

\item{timeout}{How long to wait for the element to exist.}
}
\value{
\code{x}, invisibly
}
\description{
Scrolls to an HTML element.
}
\examples{
session <- mock_selenider_session()

s(".class1") |>
  scroll_to()

}
\seealso{
Other actions: 
\code{\link{click}()},
\code{\link{hover}()},
\code{\link{set_value}()},
\code{\link{submit}()}
}
\concept{actions}
