% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenider-package.R
\docType{package}
\name{selenider-package}
\alias{selenider}
\alias{selenider-package}
\title{selenider: Concise, Lazy and Reliable Wrapper for 'chromote' and 'RSelenium'}
\description{
A user-friendly wrapper for web automation, using either 'chromote' or 'RSelenium'. Provides a simple and consistent API to make web scraping and testing scripts easy to write and understand. Elements are lazy, and automatically wait for the website to be valid, resulting in reliable and reproducible code, with no visible impact on the experience of the programmer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ashbythorpe/selenider}
  \item \url{https://ashbythorpe.github.io/selenider/}
  \item Report bugs at \url{https://github.com/ashbythorpe/selenider/issues}
}

}
\author{
\strong{Maintainer}: Ashby Thorpe \email{ashbythorpe@gmail.com} (\href{https://orcid.org/0000-0003-3106-099X}{ORCID}) [copyright holder]

}
\keyword{internal}
