% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selenider_cleanup}
\alias{selenider_cleanup}
\title{Clean up after an example}
\usage{
selenider_cleanup(env = rlang::caller_env())
}
\arguments{
\item{env}{The environment in which deferred events are contained.}
}
\value{
Nothing
}
\description{
Clean up after a selenider example, making sure all deferred events are run.
}
\keyword{internal}
