% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{selenider_session}
\alias{selenider_session}
\title{Start a session}
\usage{
selenider_session(
  browser = c("chrome", "firefox", "phantomjs", "internet explorer"),
  timeout = 4,
  driver = NULL,
  global = TRUE
)
}
\arguments{
\item{browser}{The name of the browser to run the session in.}

\item{timeout}{The default time to wait when collecting an element.}

\item{driver}{A driver object to use instead of creating one manually.}

\item{global}{Whether to set the session as the global session object.}
}
\value{
A \code{selenider_session} object.
}
\description{
Begin a session in selenider, setting the session globally unless otherwise
specified.
}
\details{
\code{selenider_session()} uses \code{\link[RSelenium:rsDriver]{RSelenium::rsDriver()}} to create a browser
session.
}
\seealso{
\code{\link[=close_session]{close_session()}}
}
