% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{set_value}
\alias{set_value}
\alias{send_keys}
\alias{clear_value}
\title{Set the value of an input}
\usage{
set_value(x, text, timeout = NULL)

send_keys(x, ..., modifiers = NULL, timeout = NULL)

clear_value(x, timeout = NULL)
}
\arguments{
\item{x}{A \code{selenider_element} object. For \code{\link[=send_keys]{send_keys()}}, this can be \code{NULL},
meaning that the keys will be sent to the current page instead of a specific
element.}

\item{text}{A string to set the value of the input element to.}

\item{timeout}{How long to wait for the element to exist.}

\item{...}{A set of inputs to send to \code{x}.}

\item{modifiers}{A character vector; one or more of "shift", "ctrl"/"control", "alt", and "command"/meta".
Note that when using chromote as a backend, these do not work on Mac.}
}
\value{
\code{x}, invisibly
}
\description{
\code{set_value()} sets the value of an HTML input element to a string.
}
\details{
\code{send_keys()} sends a set of inputs to an element.

\code{clear_value()} sets the value of an HTML element to \code{""}, removing any
existing content.
}
\examples{
html <- "
<input type='text' oninput='recordChange(event)' onkeypress='return checkEnter(event);'>
<p></p>
"

js <- "
function recordChange(e) {
  document.getElementsByTagName('p').item(0).innerText = e.target.value;
}

function checkEnter(e) {
  // If the key pressed was Enter
  if (e.keyCode == 13) {
    document.getElementsByTagName('p').item(0).innerText = 'Enter pressed!';
    return false;
  }
  return true;
}
"

session <- minimal_selenider_session(html, js = js)

html_expect(s("p"), has_exact_text(""))

input <- s("input")

set_value(input, "my text")

html_expect(s("p"), has_text("my text"))

clear_value(input)

html_expect(s("p"), has_exact_text(""))

send_keys(input, keys$enter)

html_expect(s("p"), has_text("Enter pressed!"))

}
\seealso{
Other actions: 
\code{\link{click}()},
\code{\link{hover}()},
\code{\link{scroll_to}()},
\code{\link{submit}()}
}
\concept{actions}
